move methods from InvocationExtensions into other namespaces
