+++ src/System.CommandLine.Tests/Help/HelpBuilderTests.cs
@@ -4,8 +4,8 @@
 using System.Collections.Generic;
 using System.CommandLine.Builder;
 using System.CommandLine.Help;
-using System.CommandLine.Invocation;
 using System.CommandLine.IO;
+using System.CommandLine.Parsing;
 using System.IO;
 using FluentAssertions;
 using System.Linq;
