+++ src/System.CommandLine.Tests/Invocation/CancelOnProcessTerminationTests.cs
@@ -3,6 +3,7 @@
 
 using System.CommandLine.Builder;
 using System.CommandLine.Invocation;
+using System.CommandLine.Parsing;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
 using System.Threading;
